+++ packages/front/svelte/resources/versements/versements.port.js
@@ -1,5 +1,6 @@
 import axios from "axios";
-import { flatenProviderValue } from "../../helpers/dataHelper"; // Voir si il ne vaux mieux pas passer ça en adapteur
+// Voir si il ne vaux mieux pas passer ça en adapteur
+import { flatenProviderValue } from "../../helpers/providerValueHelper";
 
 class VersementsPort {
     getEtablissementVersements(identifier) {
