+++ packages/front/svelte/views/home/home.service.js
@@ -1,8 +1,8 @@
 import axios from "axios";
 import { isRna, isStartOfSiret } from "../../helpers/validatorHelper";
 import { siretToSiren } from "../../helpers/sirenHelper";
-import { flatenProviderValue } from "../../helpers/dataHelper";
 import InteruptSearchError from "./error/InteruptSearchError";
+import { flatenProviderValue } from "../../helpers/providerValueHelper";
 
 export class HomeService {
     constructor() {
